refactor: extract filename/display modules from utils, decompose the download stream loop
- move filename sanitization to filename.rs and terminal output to
  display.rs; utils keeps the cookie/symlink/URL helpers
- split download_file_content into named helpers (resume_request,
  handle_failed_status, prefix_trusted, progress_total,
  assess_streamed_body) with behaviour preserved
- download_files takes Vec<DownloadTask> instead of a generic
  IntoIterator that was collected immediately
- drop the redundant #[allow(dead_code)] on XmlFile (pub library
  fields) and dedupe the MD5 comparison and rate formatting
- update the AGENTS.md architecture listing
